# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tutor4all API",
        "description": "This is tutor4all api",
        "termsOfService": "",
        "contact": {
            "email": "marslanali@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "#"
        },
        "version": "1.0.0"
    },
    "host": "tutor4all-api.shayansolutions.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/get-students": {
            "get": {
                "summary": "Get user",
                "description": "",
                "operationId": "getUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "header",
                        "description": "Roleid",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's List"
                    },
                    "422": {
                        "description": "Un Processable Entry"
                    }
                }
            }
        },
...
        "/get-profile": {
            "get": {
                "summary": "Get user profile",
                "description": "",
                "operationId": "getUserProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's Profile"
                    },
                    "422": {
                        "description": "Un Processable Entry"
                    }
                }
            }
        },
...
        "/update-student-profile": {
            "post": {
                "summary": "Update Student Profile",
                "description": "",
                "operationId": "updateStudentProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "First Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "Last Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fatherName",
                        "in": "query",
                        "description": "Father Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "gender_id",
                        "in": "query",
                        "description": "i.e. Male or Female",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Mobile Number",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "profileImage",
                        "in": "formData",
                        "description": "Profile Image",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "qualification",
                        "in": "query",
                        "description": "Qualification",
                        "required": false,
                        "type": "string"
                    }

                ],
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    },
                    "422": {
                        "description": "Unable to update profile"
                    }
                }
            }
        },
        "/update-tutor-profile": {
            "post": {
                "summary": "Update Tutor Profile",
                "description": "",
                "operationId": "updateTutorProfile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "First Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "Last Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "fatherName",
                        "in": "query",
                        "description": "Father Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "gender_id",
                        "in": "query",
                        "description": "i.e. Male or Female",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Mobile Number",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "profileImage",
                        "in": "formData",
                        "description": "Profile Image",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "tutor_id",
                        "in": "query",
                        "description": "Tutor id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "experience",
                        "in": "query",
                        "description": "Tutor Experience",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "qualification",
                        "in": "query",
                        "description": "Tutor qualification",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "programme_id",
                        "in": "query",
                        "description": "Tutor Programme Id",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "subject_id",
                        "in": "query",
                        "description": "Tutor Subject Id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "cnic_no",
                        "in": "query",
                        "description": "CNIC Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address",
                        "required": true,
                        "type": "string"
                    }

                ],
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    },
                    "422": {
                        "description": "Unable to update profile"
                    }
                }
            }
        },
...
    },
...
}